+++ src/addons/display_commander/continuous_monitoring.cpp
@@ -22,6 +22,7 @@
 #include "ui/new_ui/hotkeys_tab.hpp"
 #include "ui/new_ui/swapchain_tab.hpp"
 #include "utils/detour_call_tracker.hpp"
+#include "utils/taskbar_helper.hpp"
 #include "utils/display_commander_logger.hpp"
 #include "utils/logging.hpp"
 #include "utils/overlay_window_detector.hpp"
@@ -833,6 +834,13 @@ void ContinuousMonitoringThread() {
                     settings::g_mainTabSettings.adhd_multi_monitor_enabled.GetValue());
             }
 
+            // Windows taskbar visibility: 0 = no change, 1 = hide when in foreground, 2 = always hide
+            {
+                const bool in_foreground = !g_app_in_background.load(std::memory_order_acquire);
+                const int taskbar_mode = settings::g_mainTabSettings.taskbar_hide_mode.GetValue();
+                display_commander::utils::UpdateTaskbarVisibility(in_foreground, taskbar_mode);
+            }
+
             // Update keyboard tracking system
             display_commanderhooks::keyboard_tracker::Update();
 
@@ -850,6 +858,12 @@ void ContinuousMonitoringThread() {
             CALL_GUARD(utils::get_now_ns());
             last_1s_update_ns = now_ns;
             check_is_background();
+            // Recheck taskbar visibility whenever we update background state (same as in high-freq path)
+            {
+                const bool in_foreground = !g_app_in_background.load(std::memory_order_acquire);
+                const int taskbar_mode = settings::g_mainTabSettings.taskbar_hide_mode.GetValue();
+                display_commander::utils::UpdateTaskbarVisibility(in_foreground, taskbar_mode);
+            }
             g_continuous_monitoring_section.store("every1s_tasks", std::memory_order_release);
             every1s_tasks();